+++ HappyFishJump/CHANGELOG.md
@@ -1,2 +1,6 @@
-﻿# Ver 1.0.1
+﻿# Ver 1.1.0
+ - fish randomly jump in any body of water big enough
+ - sound option to disable the splash sound
+
+# Ver 1.0.1
  - added GMCM support